+++ apps/extension/src/contentscript/multitable-panel/components/modals-confirm.tsx
@@ -2,13 +2,8 @@ import React, { FC, useCallback, useEffect, useMemo, useState } from 'react'
 import BsButton from 'react-bootstrap/Button'
 import BsSpinner from 'react-bootstrap/Spinner'
 import styled from 'styled-components'
-import {
-  MutationCreateDto,
-  MutationDto,
-  useCreateMutation,
-  useEditMutation,
-  useMutableWeb,
-} from '@mweb/engine'
+import { useCreateMutation, useEditMutation, useMutableWeb } from '@mweb/engine'
+import { MutationCreateDto, MutationDto } from '@mweb/backend'
 import { Image } from './image'
 import { useEscape } from '../../hooks/use-escape'
 import { Alert, AlertProps } from './alert'
