+++ libs/backend/package.json
@@ -14,36 +14,19 @@
     "@near-wallet-selector/core": "^8.9.1",
     "@types/big.js": "^6.2.2",
     "@types/bn.js": "^5.1.5",
-    "@types/react": "^18.2.21",
-    "@types/react-dom": "^18.2.18",
-    "@types/styled-components": "^5.1.34",
     "base64-js": "^1.5.1",
-    "csstype": "^3.1.3",
     "fake-indexeddb": "^5.0.2",
     "ieee754": "^1.2.1",
-    "near-api-js": "2.1.3",
-    "near-social-vm": "github:dapplets/near-social-vm",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.5.0",
-    "react-dom": "^18.2.0",
-    "styled-components": "^5.3.6"
+    "near-api-js": "2.1.3"
   },
   "peerDependencies": {
-    "near-api-js": "2.1.3",
-    "near-social-vm": "github:dapplets/near-social-vm",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.5.0",
-    "react-dom": "^18.2.0",
-    "styled-components": "^5.3.6"
+    "near-api-js": "2.1.3"
   },
   "dependencies": {
-    "@ant-design/cssinjs": "^1.21.0",
     "@mweb/core": "workspace:*",
     "@mweb/react": "workspace:*",
-    "antd": "^5.18.3",
     "big.js": "^6.2.1",
     "caching-decorator": "^1.0.3",
-    "ethereum-blockies-base64": "^1.0.2",
     "js-sha256": "^0.11.0",
     "json-stringify-deterministic": "^1.0.12",
     "reflect-metadata": "^0.2.2"
