+++ libs/backend/src/engine.ts
@@ -1,26 +1,26 @@
 import { WalletSelector } from '@near-wallet-selector/core'
-import { getNearConfig } from './constants'
-import { NearSigner } from './app/services/near-signer/near-signer.service'
-import { SocialDbService } from './app/services/social-db/social-db.service'
-import { IStorage } from './app/services/local-db/local-storage'
-import { LocalDbService } from './app/services/local-db/local-db.service'
-import { LocalStorage } from './app/services/local-db/local-storage'
-import { MutationRepository } from './app/services/mutation/mutation.repository'
-import { ApplicationRepository } from './app/services/application/application.repository'
-import { UserLinkRepository } from './app/services/user-link/user-link.repository'
-import { ParserConfigRepository } from './app/services/parser-config/parser-config.repository'
-import { DocumentRepository } from './app/services/document/document.repository'
-import { NotificationRepository } from './app/services/notification/notification.repository'
-import { MutationService } from './app/services/mutation/mutation.service'
-import { ApplicationService } from './app/services/application/application.service'
-import { UserLinkService } from './app/services/user-link/user-link.service'
-import { ParserConfigService } from './app/services/parser-config/parser-config.service'
-import { LinkDbService } from './app/services/link-db/link-db.service'
-import { DocumentSerivce } from './app/services/document/document.service'
-import { LinkDbRepository } from './app/services/link-db/link-db.repository'
-import { UnitOfWorkService } from './app/services/unit-of-work/unit-of-work.service'
-import { NotificationService } from './app/services/notification/notification.service'
-import { ResolutionRepository } from './app/services/notification/resolution.repository'
+import { getNearConfig } from './config'
+import { NearSigner } from './services/near-signer/near-signer.service'
+import { SocialDbService } from './services/social-db/social-db.service'
+import { IStorage } from './services/local-db/local-storage'
+import { LocalDbService } from './services/local-db/local-db.service'
+import { LocalStorage } from './services/local-db/local-storage'
+import { MutationRepository } from './services/mutation/mutation.repository'
+import { ApplicationRepository } from './services/application/application.repository'
+import { UserLinkRepository } from './services/user-link/user-link.repository'
+import { ParserConfigRepository } from './services/parser-config/parser-config.repository'
+import { DocumentRepository } from './services/document/document.repository'
+import { NotificationRepository } from './services/notification/notification.repository'
+import { MutationService } from './services/mutation/mutation.service'
+import { ApplicationService } from './services/application/application.service'
+import { UserLinkService } from './services/user-link/user-link.service'
+import { ParserConfigService } from './services/parser-config/parser-config.service'
+import { LinkDbService } from './services/link-db/link-db.service'
+import { DocumentSerivce } from './services/document/document.service'
+import { LinkDbRepository } from './services/link-db/link-db.repository'
+import { UnitOfWorkService } from './services/unit-of-work/unit-of-work.service'
+import { NotificationService } from './services/notification/notification.service'
+import { ResolutionRepository } from './services/notification/resolution.repository'
 
 export type EngineConfig = {
   networkId: string
