+++ libs/backend/src/services/link-db/link-db.service.ts
@@ -1,4 +1,4 @@
-import { TransferableContext } from '../../common/transferable-context'
+import { TransferableContext } from '../../types/transferable-context'
 import { AppId } from '../application/application.entity'
 import { MutationId } from '../mutation/mutation.entity'
 import { LinkIndexRules, IndexObject, LinkedDataByAccountDto, CtxLink } from './link-db.entity'
