+++ libs/backend/src/services/near-signer/near-signer.service.ts
@@ -4,7 +4,7 @@ import { QueryResponseKind } from 'near-api-js/lib/providers/provider'
 import { KeyStorage } from '../social-db/key-storage'
 import { LocalDbService } from '../local-db/local-db.service'
 import Big from 'big.js'
-import { NearConfig } from '../../../constants'
+import { NearConfig } from '../../config'
 import { TypedError } from 'near-api-js/lib/providers'
 import BN from 'bn.js'
 import { Cacheable } from 'caching-decorator'
