+++ libs/backend/src/types-patch.d.ts
@@ -1,0 +1,1 @@
+/// <reference types="reflect-metadata" />
