+++ libs/engine/src/app/contexts/mutable-web-context/use-applications.ts
@@ -1,5 +1,5 @@
-import { ApplicationDto } from '../../services/application/dtos/application.dto'
-import { Engine } from '../../../engine'
+import { ApplicationDto } from '@mweb/backend'
+import { Engine } from '@mweb/backend'
 import { useQueryArray } from '../../hooks/use-query-array'
 
 export const useApplications = (engine: Engine) => {
