+++ libs/engine/src/index.ts
@@ -1,15 +1,4 @@
 export * as customElements from './custom-elements'
-export { MutationWithSettings } from './app/services/mutation/mutation.entity'
-export { MutationDto } from './app/services/mutation/dtos/mutation.dto'
-export { MutationCreateDto } from './app/services/mutation/dtos/mutation-create.dto'
-export {
-  AppWithSettings,
-  AppInstanceWithSettings,
-} from './app/services/application/application.entity'
-export { ApplicationDto } from './app/services/application/dtos/application.dto'
-export { ApplicationCreateDto } from './app/services/application/dtos/application-create.dto'
-export { LocalStorage } from './app/services/local-db/local-storage'
-export { IStorage } from './app/services/local-db/local-storage'
 export { App } from './app/app'
 export { useEngine } from './app/contexts/engine-context'
 export {
@@ -19,11 +8,8 @@ export {
   useMutationApp,
 } from './app/contexts/mutable-web-context'
 export { ShadowDomWrapper } from './app/components/shadow-dom-wrapper'
-export { EngineConfig } from './engine'
 export { App as MutableWebProvider } from './app/app'
 export { useAppDocuments } from './app/contexts/mutable-web-context/use-app-documents'
-export { DocumentDto } from './app/services/document/dtos/document.dto'
-export { DocumentCreateDto } from './app/services/document/dtos/document-create.dto'
 export {
   NotificationProvider,
   useNotifications,
@@ -33,5 +19,3 @@ export {
   useViewNotification,
   useHideNotification,
 } from './app/contexts/notification-context'
-export { NotificationType } from './app/services/notification/notification.entity'
-export { NotificationDto } from './app/services/notification/dtos/notification.dto'
